reactored/splitted tests
